@@ -3,4 +3,5 @@
 }
 <h1>@ViewData["Title"]</h1>
 
-<p>Use this page to detail your site's privacy policy.</p>
+<p>This site makes use of an autogenerated cert for HTTPS://localhost/.</p>
+<p>Should this cause problems raise and issue on GitHub.</p>
